fix some bug fix #33
